+++ Qtica/utils/methods.py
@@ -1,4 +1,5 @@
 import unicodedata
+from PySide6.QtCore import Qt
 
 
 def slugify(original: str) -> str:
@@ -23,3 +24,13 @@ def slugify(original: str) -> str:
     slugified = slugified.replace(" ", "-")
 
     return slugified
+
+
+def qt_corner_to_edge(corner: Qt.Corner,
+                      default: Qt.Edge = Qt.Edge(0)) -> Qt.Edge:
+    return {
+        Qt.Corner.TopLeftCorner: Qt.Edge.LeftEdge | Qt.Edge.TopEdge,
+        Qt.Corner.TopRightCorner: Qt.Edge.RightEdge | Qt.Edge.TopEdge,
+        Qt.Corner.BottomLeftCorner: Qt.Edge.LeftEdge | Qt.Edge.BottomEdge,
+        Qt.Corner.BottomRightCorner: Qt.Edge.RightEdge | Qt.Edge.BottomEdge
+    }.get(corner, default)