+++ Qtica/widgets/dialogs/__init__.py
@@ -1,2 +1,3 @@
 from .display_large_text import LargTextDialog
 from .silent_mode import SilentTextDialog
+from .mask_dialog import MaskDialog