+++ Qtica/widgets/stacked_widget/__init__.py
@@ -1,1 +1,2 @@
 from .sliding import SlidingStackedWidget
+from .stacked_widget import StackedWidget